refactor: move classes to correct packages
