+++ common/src/main/java/com/wynntils/screens/maps/managers/CategoryManagementScreen.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2026.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.screens.maps;
+package com.wynntils.screens.maps.managers;
 
 import com.google.common.collect.Lists;
 import com.wynntils.core.components.Models;
@@ -10,32 +10,30 @@ import com.wynntils.core.components.Services;
 import com.wynntils.core.consumers.screens.WynntilsScreen;
 import com.wynntils.screens.base.TooltipProvider;
 import com.wynntils.screens.base.widgets.TextInputBoxWidget;
-import com.wynntils.services.mapdata.attributes.DefaultMapAttributes;
-import com.wynntils.screens.maps.categorymanagerwidgets.CategorySearchWidget;
-import com.wynntils.screens.maps.categorymanagerwidgets.CategoryTreeWidget;
-import com.wynntils.screens.maps.categorymanagerwidgets.DeleteButtonWidget;
-import com.wynntils.screens.maps.categorymanagerwidgets.ResetButtonWidget;
-import com.wynntils.screens.maps.categorymanagerwidgets.OptionsScrollBarWidget;
-import com.wynntils.screens.maps.categorymanagerwidgets.OverrideSelectionWidget;
-import com.wynntils.screens.maps.categorymanagerwidgets.SaveButtonWidget;
-import com.wynntils.screens.maps.categorymanagerwidgets.optionwidgets.AbstractOptionWidget;
-import com.wynntils.screens.maps.categorymanagerwidgets.optionwidgets.ColorOptionWidget;
-import com.wynntils.screens.maps.categorymanagerwidgets.optionwidgets.IconOptionWidget;
-import com.wynntils.screens.maps.categorymanagerwidgets.optionwidgets.FloatSliderOptionWidget;
-import com.wynntils.screens.maps.categorymanagerwidgets.optionwidgets.IntSliderOptionWidget;
-import com.wynntils.screens.maps.categorymanagerwidgets.optionwidgets.TextOptionWidget;
-import com.wynntils.screens.maps.categorymanagerwidgets.optionwidgets.TextShadowOptionWidget;
-import com.wynntils.screens.maps.categorymanagerwidgets.optionwidgets.ToggleOptionWidget;
-import com.wynntils.screens.maps.type.OptionCategory;
-import com.wynntils.screens.maps.type.OverrideType;
+import com.wynntils.screens.maps.MainMapScreen;
+import com.wynntils.screens.maps.managers.widgets.CategorySearchWidget;
+import com.wynntils.screens.maps.managers.widgets.CategoryTreeWidget;
+import com.wynntils.screens.maps.managers.widgets.DeleteButtonWidget;
+import com.wynntils.screens.maps.managers.widgets.ResetButtonWidget;
+import com.wynntils.screens.maps.managers.widgets.OptionsScrollBarWidget;
+import com.wynntils.screens.maps.managers.widgets.OverrideSelectionWidget;
+import com.wynntils.screens.maps.managers.widgets.SaveButtonWidget;
+import com.wynntils.screens.maps.managers.widgets.options.AbstractOptionWidget;
+import com.wynntils.screens.maps.managers.widgets.options.ColorOptionWidget;
+import com.wynntils.screens.maps.managers.widgets.options.IconOptionWidget;
+import com.wynntils.screens.maps.managers.widgets.options.FloatSliderOptionWidget;
+import com.wynntils.screens.maps.managers.widgets.options.IntSliderOptionWidget;
+import com.wynntils.screens.maps.managers.widgets.options.TextOptionWidget;
+import com.wynntils.screens.maps.managers.widgets.options.TextShadowOptionWidget;
+import com.wynntils.screens.maps.managers.widgets.options.ToggleOptionWidget;
+import com.wynntils.screens.maps.managers.type.OptionCategory;
+import com.wynntils.screens.maps.managers.type.OverrideType;
 import com.wynntils.services.mapdata.attributes.type.MapAttributes;
 import com.wynntils.services.mapdata.attributes.type.MapMarkerOptions;
 import com.wynntils.services.mapdata.attributes.type.MapVisibility;
-import com.wynntils.utils.colors.CommonColors;
 import com.wynntils.utils.mc.McUtils;
 import com.wynntils.utils.render.RenderUtils;
 import com.wynntils.utils.render.Texture;
-import com.wynntils.utils.render.type.TextShadow;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.events.GuiEventListener;
 import net.minecraft.client.gui.screens.Screen;
