+++ common/src/main/java/com/wynntils/screens/maps/managers/type/CategoryTree.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2026.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.screens.maps.type;
+package com.wynntils.screens.maps.managers.type;
 
 import com.wynntils.utils.StringUtils;
 import java.util.ArrayList;
