+++ common/src/main/java/com/wynntils/screens/maps/managers/widgets/CategorySearchWidget.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2026.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.screens.maps.categorymanagerwidgets;
+package com.wynntils.screens.maps.managers.widgets;
 
 import com.wynntils.core.text.StyledText;
 import com.wynntils.screens.base.TextboxScreen;
