+++ common/src/main/java/com/wynntils/screens/maps/managers/widgets/DoubleScrollBarWidget.java
@@ -2,10 +2,10 @@
  * Copyright © Wynntils 2026.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.screens.maps.categorymanagerwidgets;
+package com.wynntils.screens.maps.managers.widgets;
 
 import com.mojang.blaze3d.platform.cursor.CursorTypes;
-import com.wynntils.screens.maps.CategoryManagementScreen;
+import com.wynntils.screens.maps.managers.CategoryManagementScreen;
 import com.wynntils.utils.MathUtils;
 import com.wynntils.utils.render.RenderUtils;
 import com.wynntils.utils.render.Texture;
